requirements.txt
pandas
slack-sdk
mysql-connector-python
python-dotenv
SQLAlchemy==1.4.16